Refactor :hammer: Extract components to separate files
